# swagger.yaml
openapi: 3.0.0
info:
  title: ZUCT
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /course/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Course
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Course
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    post:
      summary: CRUD ACTION update
      tags:
        - Course
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
  /course/findByenrol/{key}:
    get:
      summary: CRUD ACTION findByenrol
      tags:
        - Course
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /course/findBytutor/{key}:
    get:
      summary: CRUD ACTION findBytutor
      tags:
        - Course
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    
  #-----  Enrollment  -----
...
  /enrollment/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Enrollment
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Enrollment
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
    post:
      summary: CRUD ACTION update
      tags:
        - Enrollment
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Enrollment"
  /enrollment/findByparticipate/{key}:
    get:
      summary: CRUD ACTION findByparticipate
      tags:
        - Enrollment
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
  /enrollment/findBytutotr/{key}:
    get:
      summary: CRUD ACTION findBytutotr
      tags:
        - Enrollment
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    
  #-----  LECTURE  -----
...
  /lecture/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - LECTURE
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - LECTURE
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LECTURE"
    post:
      summary: CRUD ACTION update
      tags:
        - LECTURE
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LECTURE"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LECTURE"
  /lecture/findByparticipate/{key}:
    get:
      summary: CRUD ACTION findByparticipate
      tags:
        - LECTURE
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LECTURE"
  /lecture/findBytutor/{key}:
    get:
      summary: CRUD ACTION findBytutor
      tags:
        - LECTURE
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LECTURE"
    
  #-----  STUDENT  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Course:
      properties:
        _id:
          type: string
        C_lecture:
          type: string
        Course_ID:
          type: number
        c_name:
          type: string
      required:
        -  _id
        - Course_ID
    Enrollment:
      properties:
        _id:
          type: string
        Nationality:
          type: string
        Qualification:
          type: string
      required:
        -  _id
    LECTURE:
      properties:
        _id:
          type: string
        Course:
          type: string
        Department:
          type: string
        First_name:
          type: string
        Last_name:
          type: string
      required:
        -  _id
...